change money values from BigInteger to BigDecimal and fixes to offset payment method

+++ src/main/java/com/example/ethereumserviceapp/model/Case.java
@@ -1,6 +1,6 @@
 package com.example.ethereumserviceapp.model;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -21,5 +21,5 @@ public class Case {
     private State state;
     private LinkedHashMap<LocalDateTime, State> history;
     private List<CasePayment> paymentHistory;
-    private BigInteger offset;
+    private BigDecimal offset;
 }
